Cоздал компоненту offer-block
